Form Class: move interfaces to their own files
